Update BinarySearch.java
What is the Binary Search?

Binary Search is a search a sorted array by repeatedly dividing the search interval in half. Begin with an interval covering the whole array.
If the value of the search key is less than the item in the middle of the interval, narrow the interval to the lower half. 
Otherwise, narrow it to the upper half.
It will check until the value is found or the interval is empty.

This code will return -1 if there is no target number.